% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_value.R
\name{critical_value}
\alias{critical_value}
\title{Calculate test critical value}
\usage{
critical_value(p_1_hat, n_1, p_0_hat, n_0, alpha, s, h, gamma)
}
\arguments{
\item{p_1_hat}{estimated population 1 p parameter}

\item{n_1}{population 1 sample size}

\item{p_0_hat}{estimated population 0 p parameter}

\item{n_0}{population 0 sample size}

\item{alpha}{test significance level}

\item{s}{either 1 (for one sided test) or 2 (for two sided test)}

\item{h}{number of hypothesis tested in the same experiment (for Bonferroni correction)}

\item{gamma}{minimum required lift}
}
\value{
critical value C
}
\description{
this function calculates the test statistic critical
value C which guarantees significance level (false discovery rate) alpha
}
\details{
when doing one-sided tests it's usually the case
that population 1 is considered the treatment and
population 0 serves as the control. In 2 sided tests each
usually represents a different treatment.
}
\examples{
library(hype)

# For a thorugh simulation validation see
# https://github.com/IyarLin/hype/blob/master/inst/variuos_results_for_hypothesis_testing.pdf

# below we'll generate critical values for 2 hypothesis tests performed
# in the same experiemnt
# note that h is set to 2 in order to reflect that

## hypothesis test number 1
critical_value(
  p_1_hat = 0.212, n_1 = 10000, p_0_hat = 0.219, n_0 = 8000,
  alpha = 0.05, s = 1, h = 2, gamma = 0.01
)

## hypothesis test number 2
critical_value(
  p_1_hat = 0.11, n_1 = 5000, p_0_hat = 0.132, n_0 = 3000,
  alpha = 0.05, s = 2, h = 2, gamma = 0
)
}
