% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_sample_size.R
\name{min_sample_size}
\alias{min_sample_size}
\title{Calculate minimum required sample size}
\usage{
min_sample_size(p_1, p_0, alpha, s, h, gamma, power)
}
\arguments{
\item{p_1}{assumed population 1 p parameter (can also be thought of as the MDE)}

\item{p_0}{assumed population 0 p parameter}

\item{alpha}{test significance level}

\item{s}{either 1 (for one sided test) or 2 (for two sided test)}

\item{h}{number of hypothesis tested in the same experiment (for Bonferroni correction)}

\item{gamma}{minimum required lift}

\item{power}{desired test power (inverse of type 2 error)}

\item{n_1}{population 0 sample size}
}
\value{
minimum requires sample size per treatment group
}
\description{
this function calculates the minimum
required sample size (equal for both treatment groups)
which acheives a desired MDE and power provided alpha, s,
h and gamma.
}
\details{
when doing one-sided tests it's usually the case
that population 1 is considered the treatment and population
population 0 serves as the control. In 2 sided tests each
usually represents a different treatment. 2 sided tests with
non zero gamma aren't supported yet.
}
\examples{
library(hype)

# For a thorugh simulation validation see
# https://github.com/IyarLin/hype/blob/master/inst/variuos_results_for_hypothesis_testing.pdf

# below we'll calculate minimum requires sample size per group
# for 2 hypothesis tests performed
# in the same experiemnt
# note that h is set to 2 in order to reflect that

## hypothesis test number 1
min_sample_size(p_1 = 0.225, p_0 = 0.2,
      alpha = 0.05, s = 1, h = 2, gamma = 0.01,
      power = 0.8)

## hypothesis test number 2
min_sample_size(p_1 = 0.118, p_0 = 0.1,
      alpha = 0.05, s = 2, h = 2, gamma = 0,
      power = 0.8)
}
