% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\title{Calculate test power}
\usage{
power(p_1, n_1, p_0, n_0, alpha, s, h, gamma)
}
\arguments{
\item{p_1}{assumed population 1 p parameter}

\item{n_1}{population 1 sample size}

\item{p_0}{assumed population 0 p parameter}

\item{n_0}{population 0 sample size}

\item{alpha}{test significance level}

\item{s}{either 1 (for one sided test) or 2 (for two sided test)}

\item{h}{number of hypothesis tested in the same experiment (for Bonferroni correction)}

\item{gamma}{minimum required lift}
}
\value{
test power
}
\description{
this function calculates the test power
(inverse of the type 2 probability).
}
\details{
when doing one-sided tests it's usually the case
that population 1 is considered the treatment and
population 0 serves as the control. In 2 sided tests each
usually represents a different treatment. 2 sided tests with
non zero gamma aren't supported yet.
}
\examples{
library(hype)

# For a thorugh simulation validation see
# https://github.com/IyarLin/hype/blob/master/inst/variuos_results_for_hypothesis_testing.pdf

# below we'll calculate power for 2 hypothesis tests performed
# in the same experiemnt
# note that h is set to 2 in order to reflect that

## hypothesis test number 1
power(p_1 = 0.225, n_1 = 10000, p_0 = 0.2, n_0 = 8000,
  alpha = 0.05, s = 1, h = 2, gamma = 0.01)

## hypothesis test number 2
power(p_1 = 0.118, n_1 = 5000, p_0 = 0.1, n_0 = 3000,
  alpha = 0.05, s = 2, h = 2, gamma = 0)
}
