% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDE.R
\name{MDE}
\alias{MDE}
\title{Calculate test Minimum detectable effect (MDE)}
\usage{
MDE(power, n_1, p_0, n_0, alpha, s, h, gamma)
}
\arguments{
\item{power}{desired test power (inverse of type 2 error)}

\item{n_1}{population 1 sample size}

\item{p_0}{assumed population 0 p parameter}

\item{n_0}{population 0 sample size}

\item{alpha}{test significance level}

\item{s}{either 1 (for one sided test) or 2 (for two sided test)}

\item{h}{number of hypothesis tested in the same experiment (for Bonferroni correction)}

\item{gamma}{minimum required lift}
}
\value{
test MDE
}
\description{
this function calculates the test minimum
dedtectable effect (MDE). This is the smallest effect
size p_1 - p_0 which yields the desired power and provided
sample sizes, alpha, s, h and gamma.
}
\details{
In 2 sided tests each
usually represents a different treatment.
When doing one-sided tests it's usually the case
that population 1 is considered the treatment and
population 0 serves as the control. At any rate,
one sided tests are always of the form p_1 - p_0 > C.
For that reason for 1 sided tests one must set: gamma >= 0
}
\examples{
library(hype)

# For a thorugh simulation validation see
# https://github.com/IyarLin/hype/blob/master/inst/variuos_results_for_hypothesis_testing.pdf

# below we'll calculate MDE for 2 hypothesis tests performed
# in the same experiemnt
# note that h is set to 2 in order to reflect that

## hypothesis test number 1
MDE(
  n_1 = 10000, p_0 = 0.2, n_0 = 8000,
  alpha = 0.05, s = 1, h = 2, gamma = 0.01, power = 0.8
)

## hypothesis test number 2
MDE(
  n_1 = 5000, p_0 = 0.1, n_0 = 3000,
  alpha = 0.05, s = 2, h = 2, gamma = 0, power = 0.8
)
}
